Use Failure instead of exception in ProblemEvents
