+++ platforms/ide/tooling-api-builders/src/main/java/org/gradle/tooling/internal/provider/runner/ProblemsProgressEventConsumer.java
@@ -31,6 +31,7 @@ import org.gradle.api.problems.internal.TaskPathLocation;
 import org.gradle.internal.build.event.types.DefaultAdditionalData;
 import org.gradle.internal.build.event.types.DefaultDetails;
 import org.gradle.internal.build.event.types.DefaultDocumentationLink;
+import org.gradle.internal.build.event.types.DefaultFailure;
 import org.gradle.internal.build.event.types.DefaultLabel;
 import org.gradle.internal.build.event.types.DefaultProblemCategory;
 import org.gradle.internal.build.event.types.DefaultProblemDescriptor;
@@ -40,6 +41,7 @@ import org.gradle.internal.build.event.types.DefaultSeverity;
 import org.gradle.internal.build.event.types.DefaultSolution;
 import org.gradle.internal.operations.OperationIdentifier;
 import org.gradle.internal.operations.OperationProgressEvent;
+import org.gradle.tooling.internal.protocol.InternalFailure;
 import org.gradle.tooling.internal.protocol.InternalProblemEvent;
 import org.gradle.tooling.internal.protocol.problem.InternalAdditionalData;
 import org.gradle.tooling.internal.protocol.problem.InternalDetails;
@@ -103,9 +105,18 @@ public class ProblemsProgressEventConsumer extends ClientForwardingBuildOperatio
                 toInternalLocations(problem.getLocations()),
                 toInternalDocumentationLink(problem.getDocumentationLink()),
                 toInternalSolutions(problem.getSolutions()),
-                toInternalAdditionalData(problem.getAdditionalData())
+                toInternalAdditionalData(problem.getAdditionalData()),
+                toInternalFailure(problem.getException())
             )
         );
+    }
+
+    @Nullable
+    private static InternalFailure toInternalFailure(@Nullable RuntimeException ex) {
+        if(ex == null) {
+            return null;
+        }
+        return DefaultFailure.fromThrowable(ex);
     }
 
     private DefaultProblemDescriptor cerateDefaultProblemDescriptor(OperationIdentifier parentBuildOperationId) {
