+++ platforms/ide/tooling-api/src/main/java/org/gradle/tooling/events/problems/internal/DefaultProblemsOperationDescriptor.java
@@ -22,6 +22,7 @@ import org.gradle.tooling.events.internal.DefaultOperationDescriptor;
 import org.gradle.tooling.events.problems.AdditionalData;
 import org.gradle.tooling.events.problems.Details;
 import org.gradle.tooling.events.problems.DocumentationLink;
+import org.gradle.tooling.events.problems.FailureContainer;
 import org.gradle.tooling.events.problems.Label;
 import org.gradle.tooling.events.problems.Location;
 import org.gradle.tooling.events.problems.ProblemCategory;
@@ -43,8 +44,8 @@ public class DefaultProblemsOperationDescriptor extends DefaultOperationDescript
     private final DocumentationLink documentationLink;
     private final List<Solution> solutions;
     private final AdditionalData additionalData;
-//    @Nullable
-//    private final ExceptionContainer exception;
+    @Nullable
+    private final FailureContainer exception;
 
     public DefaultProblemsOperationDescriptor(
         InternalOperationDescriptor internalDescriptor,
@@ -56,9 +57,8 @@ public class DefaultProblemsOperationDescriptor extends DefaultOperationDescript
         List<Location> locations,
         @Nullable DocumentationLink documentationLink,
         List<Solution> solutions,
-        AdditionalData additionalData
-//        ,
-//        @Nullable ExceptionContainer exception
+        AdditionalData additionalData,
+        @Nullable FailureContainer exception
     ) {
         super(internalDescriptor, parent);
         this.category = category;
@@ -69,7 +69,7 @@ public class DefaultProblemsOperationDescriptor extends DefaultOperationDescript
         this.documentationLink = documentationLink;
         this.solutions = solutions;
         this.additionalData = additionalData;
-//        this.exception = exception;
+        this.exception = exception;
     }
 
     public ProblemCategory getCategory() {
@@ -111,8 +111,8 @@ public class DefaultProblemsOperationDescriptor extends DefaultOperationDescript
         return additionalData;
     }
 
-//    @Nullable
-//    public ExceptionContainer getException() {
-//        return exception;
-//    }
+    @Nullable
+    public FailureContainer getFailure() {
+        return exception;
+    }
 }
